Rename lib
