+++ include/egc.h
